SESSION MANAGEMENT

+++ Infinity-Scheduler/Infinity-Scheduler/server/deleteEvent.php
@@ -1,27 +1,29 @@
 <?php
-	require_once "config.php";
-	if(isset($_GET["args"]) && !empty(trim($_GET["args"]))){
+	require_once "session.php";
+	if (startOrResumeSession()){
+		require_once "config.php";
+		if(isset($_GET["args"]) && !empty(trim($_GET["args"]))){
 		
 
-		$args = urldecode($_GET['args']);
+			$args = urldecode($_GET['args']);
 		
-		$args = json_decode($args, true);
+			$args = json_decode($args, true);
 		
 		
 		
-		$stmt = mysqli_stmt_init($link);
-		if (mysqli_stmt_prepare($stmt, "DELETE FROM TaskEvent WHERE UID=? AND TID=?")) {
+			$stmt = mysqli_stmt_init($link);
+			if (mysqli_stmt_prepare($stmt, "DELETE FROM TaskEvent WHERE UID=? AND TID=?")) {
 
-			/* bind parameters for markers */
-			mysqli_stmt_bind_param($stmt, "ii", $args["UID"], $args["TID"]);
+				/* bind parameters for markers */
+				mysqli_stmt_bind_param($stmt, "ii", $_SESSION["UID"], $args["TID"]);
 
-			/* execute query */
-			mysqli_stmt_execute($stmt);
-			if (mysqli_stmt_affected_rows($stmt) == 1){
-				echo formatSuccess("deleted");
-			} else {
-				echo formatError("nothing deleted");
-			}
+				/* execute query */
+				mysqli_stmt_execute($stmt);
+				if (mysqli_stmt_affected_rows($stmt) == 1){
+					echo formatSuccess("deleted");
+				} else {
+					echo formatError("nothing deleted");
+				}
 			
 			
 			
@@ -31,15 +33,17 @@
 
 			
 			  
-			/* close statement */
-			mysqli_stmt_close($stmt);
+				/* close statement */
+				mysqli_stmt_close($stmt);
+
+			} else {
+				echo formatError("failure: bad query");
+			}
 
 		} else {
-			echo formatError("failure: bad query");
+			echo formatError("failure: no args");
 		}
-
-	} else {
-		echo formatError("failure: no args");
 	}
+	
 
 ?>
